Optimize Sean Penn

+++ catscripts/Entertainment/Actors_and_actresses/Sean_Penn/SeanPenn.sh
@@ -1,16 +1,20 @@
 #!/bin/bash
 
-export KEYWORDS_ACTOR_SEANPENN="Sean(| )Penn"
+KEYWORDS_ACTOR_SEANPENN="Sean(| )Penn"
 
-egrep -i "$KEYWORDS_ACTOR_SEANPENN" newpages.txt >> SeanPenn.txt
-
-SEANPENN=`stat --print=%s SeanPenn.txt`
-
-if [ $SEANPENN -ne 0 ];
+if [ "$1" == "" ];
 then
-  export CATFILE="SeanPenn.txt"
-  export CATNAME="Sean Penn"
-  $CATEGORIZE
-fi
 
-rm SeanPenn.txt
+  SEANPENN=`egrep -i "$KEYWORDS_ACTOR_SEANPENN" newpages.txt`
+
+  if [ $SEANPENN -ne 0 ];
+  then
+    egrep -i "$KEYWORDS_ACTOR_SEANPENN" newpages.txt > SeanPenn.txt
+    export CATFILE="SeanPenn.txt"
+    export CATNAME="Sean Penn"
+    $CATEGORIZE
+    rm SeanPenn.txt
+    unset SEANPENN
+  fi
+
+fi
